refactor: consolidate mass email packages under `org.shanoir.ng.email`
