+++ shanoir-ng-users/src/main/java/org/shanoir/ng/email/MassEmailApiController.java
@@ -12,13 +12,12 @@
  * along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
  */
 
-package org.shanoir.ng.massemail.controller;
+package org.shanoir.ng.email;
 
 import java.util.List;
 
-import org.shanoir.ng.massemail.model.MassEmailRequest;
-import org.shanoir.ng.massemail.model.RecipientGroup;
-import org.shanoir.ng.massemail.service.MassEmailService;
+import org.shanoir.ng.email.model.MassEmailRequest;
+import org.shanoir.ng.email.model.RecipientGroup;
 import org.shanoir.ng.shared.error.FieldErrorMap;
 import org.shanoir.ng.shared.exception.ErrorDetails;
 import org.shanoir.ng.shared.exception.ErrorModel;
