+++ shanoir-ng-users/src/main/java/org/shanoir/ng/email/MassEmailService.java
@@ -12,11 +12,11 @@
  * along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
  */
 
-package org.shanoir.ng.massemail.service;
+package org.shanoir.ng.email;
 
 import java.util.List;
 
-import org.shanoir.ng.massemail.model.RecipientGroup;
+import org.shanoir.ng.email.model.RecipientGroup;
 import org.shanoir.ng.shared.exception.SecurityException;
 import org.shanoir.ng.user.model.User;
 import org.springframework.stereotype.Service;
