+++ shanoir-ng-users/src/main/java/org/shanoir/ng/email/MassEmailServiceImpl.java
@@ -12,14 +12,13 @@
  * along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
  */
 
-package org.shanoir.ng.massemail.service;
+package org.shanoir.ng.email;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import org.shanoir.ng.email.EmailService;
-import org.shanoir.ng.massemail.model.RecipientGroup;
+import org.shanoir.ng.email.model.RecipientGroup;
 import org.shanoir.ng.shared.exception.SecurityException;
 import org.shanoir.ng.user.model.User;
 import org.shanoir.ng.user.repository.UserRepository;
